+++ packages/cli/command/src/controller/command-factory.ts
@@ -15,6 +15,7 @@ class CommandFactory {
     this.register('gen:api', new GenApiCommandHandler());
     this.register('gen:config', new GenConfigCommandHandler());
     this.register('gen:api-config', new GenApiConfigCommandHandler());
+    this.register('gen:svg-config', new GenSvgConfigCommandHandler());
   }
 
   register(command: string, handler: CommandHandler) {
@@ -84,6 +85,12 @@ class GenApiCommandHandler implements CommandHandler {
 class GenConfigCommandHandler implements CommandHandler {
   async execute(controller: Controller): Promise<void> {
     await controller.initConfig();
+  }
+}
+
+class GenSvgConfigCommandHandler implements CommandHandler {
+  async execute(controller: Controller): Promise<void> {
+    await controller.genSvgConfig();
   }
 }